Remove break in UninstallPlugins
The indexes of the Install folder registry key aren’t an unbroken number sequence - ie if the user only chooses to install into some graphics programs, but not all. So unfortunately we can’t break on a missing index number the way the code currently works. (But otherwise, this usually would have been the right thing to do though.)

Hat tip to Neil Sagitov of Lokas Software for spotting this.